LayoutTests/crypto/webkitSubtle/gc-2-expected.txt

CONSOLE MESSAGE: line 19: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test that window.crypto.webkitSubtle wrapper preserves custom properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS [object Crypto] is defined.
PASS [object WebKitSubtleCrypto] is defined.
PASS anotherWindowCrypto.webkitSubtle.foo is "bar"
PASS successfullyParsed is true

TEST COMPLETE
